refactor: extract token preview widgets
