+++ widgetbook/lib/ui/foundation/radius.dart
@@ -2,49 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:groceries_app/ui/ui.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart';
 
-class RadiusWidget extends StatelessWidget {
-  const RadiusWidget({
-    super.key,
-    required this.name,
-    required this.radius,
-  });
-
-  final double radius;
-  final String name;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(8),
-      child: Row(
-        children: [
-          SizedBox(
-            width: 100,
-            child: Text(
-              name,
-            ),
-          ),
-          Container(
-            height: 80,
-            width: 80,
-            decoration: BoxDecoration(
-              color: DesignSystemColor.brand.shade400.withOpacity(0.15),
-              borderRadius: BorderRadius.circular(radius),
-              border: Border.all(
-                width: 2,
-                color: DesignSystemColor.brand.shade400,
-              ),
-            ),
-          ),
-          const SizedBox(
-            width: 16,
-          ),
-          Text('${radius.toStringAsFixed(0)}px'),
-        ],
-      ),
-    );
-  }
-}
+import 'radius_preview.dart';
 
 @UseCase(
   name: 'Default',
@@ -56,35 +14,35 @@ Widget buildDesignSystemRadiusUseCase(BuildContext context) {
   return const SingleChildScrollView(
     child: Column(
       children: [
-        RadiusWidget(
+        RadiusPreview(
           name: 'none',
           radius: DesignSystemRadius.none,
         ),
-        RadiusWidget(
+        RadiusPreview(
           name: 'xs',
           radius: DesignSystemRadius.xs,
         ),
-        RadiusWidget(
+        RadiusPreview(
           name: 'sm',
           radius: DesignSystemRadius.sm,
         ),
-        RadiusWidget(
+        RadiusPreview(
           name: 'md',
           radius: DesignSystemRadius.md,
         ),
-        RadiusWidget(
+        RadiusPreview(
           name: 'lg',
           radius: DesignSystemRadius.lg,
         ),
-        RadiusWidget(
+        RadiusPreview(
           name: 'xl',
           radius: DesignSystemRadius.xl,
         ),
-        RadiusWidget(
+        RadiusPreview(
           name: 'xxl',
           radius: DesignSystemRadius.xxl,
         ),
-        RadiusWidget(
+        RadiusPreview(
           name: 'full',
           radius: DesignSystemRadius.full,
         ),
